# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AP Automation API",
    "description": "API for accounts payable automation with ERP-neutral extraction and configurable connectors",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard/packages/{package_id}/invoices": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Package Invoices",
        "description": "Returns all invoices in a package.",
        "operationId": "list_package_invoices_dashboard_packages__package_id__invoices_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by invoice status",
              "title": "Status"
            },
            "description": "Filter by invoice status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  },
                  "title": "Response List Package Invoices Dashboard Packages  Package Id  Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Invoice Details",
        "description": "Returns complete invoice details including line items, validation, and reconciliation.",
        "operationId": "get_package_invoice_dashboard_packages__package_id__invoices__invoice_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          },
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Invoice identifier",
              "title": "Invoice Id"
            },
            "description": "Invoice identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/packages/{package_id}/invoices": {
      "post": {
        "tags": [
          "Packages"
        ],
        "summary": "Upload Invoices",
        "description": "Upload invoice documents to the package.",
        "operationId": "upload_invoices_packages__package_id__invoices_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_invoices_packages__package_id__invoices_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Invoices Packages  Package Id  Invoices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "List Invoices",
        "description": "List invoices with optional filtering.",
        "operationId": "list_invoices_invoices_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Package Id"
            }
          },
          {
            "name": "vendor_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Code"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Get Invoice",
        "description": "Get a specific invoice by ID.",
        "operationId": "get_invoice_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Invoices"
        ],
        "summary": "Update Invoice",
        "description": "Update invoice fields.\n\nUse this to set or override GL accounts, dimensions, etc.",
        "operationId": "update_invoice_invoices__invoice_id__patch",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/lines/{line_index}": {
      "patch": {
        "tags": [
          "Invoices"
        ],
        "summary": "Update Invoice Line",
        "description": "Update a specific line item.",
        "operationId": "update_invoice_line_invoices__invoice_id__lines__line_index__patch",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          },
          {
            "name": "line_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Line Index"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceLineUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/invoices/{invoice_id}/validate": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "summary": "Validate Invoice",
        "description": "Run validation on an invoice.",
        "operationId": "validate_invoice_invoices__invoice_id__validate_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Invoice Invoices  Invoice Id  Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/post": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "summary": "Post Invoice",
        "description": "Post a single invoice to ERP.",
        "operationId": "post_invoice_invoices__invoice_id__post_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          },
          {
            "name": "connector_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "business_central",
              "title": "Connector Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post Invoice Invoices  Invoice Id  Post Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_invoices_packages__package_id__invoices_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_invoices_packages__package_id__invoices_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceDetailResponse": {
        "properties": {
          "invoice_id": {
            "type": "string",
            "title": "Invoice Id"
          },
          "lot_number": {
            "type": "string",
            "title": "Lot Number"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "feed_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feed Type"
          },
          "head_count": {
            "type": "integer",
            "title": "Head Count"
          },
          "days_on_feed": {
            "type": "integer",
            "title": "Days On Feed"
          },
          "cost_per_head": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Per Head"
          },
          "confidence": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "extracted_fields": {
            "items": {
              "$ref": "#/components/schemas/ExtractedField"
            },
            "type": "array",
            "title": "Extracted Fields"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineItem"
            },
            "type": "array",
            "title": "Line Items"
          },
          "totals": {
            "$ref": "#/components/schemas/InvoiceTotals"
          },
          "agent_commentary": {
            "items": {
              "$ref": "#/components/schemas/AgentCommentary"
            },
            "type": "array",
            "title": "Agent Commentary"
          },
          "gl_coding": {
            "items": {
              "$ref": "#/components/schemas/GLCodingEntry"
            },
            "type": "array",
            "title": "Gl Coding"
          },
          "validation_checks": {
            "items": {
              "$ref": "#/components/schemas/ValidationCheck"
            },
            "type": "array",
            "title": "Validation Checks"
          },
          "reconciliation": {
            "$ref": "#/components/schemas/ReconciliationResult"
          },
          "source_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf Url",
            "description": "URL to source PDF"
          },
          "statement_highlight_region": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement Highlight Region",
            "description": "Region to highlight in statement"
          }
        },
        "type": "object",
        "required": [
          "invoice_id",
          "lot_number",
          "amount",
          "status",
          "head_count",
          "days_on_feed",
          "cost_per_head",
          "confidence",
          "totals",
          "reconciliation"
        ],
        "title": "InvoiceDetailResponse",
        "description": "Complete invoice detail for the detail panel."
      },
...
      "InvoiceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "InvoiceListResponse",
        "description": "Paginated list of invoices."
      },
      "InvoiceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "package_id": {
            "type": "string",
            "format": "uuid",
            "title": "Package Id"
          },
          "invoice_number": {
            "type": "string",
            "title": "Invoice Number"
          },
          "invoice_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Date"
          },
          "vendor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Name"
          },
          "vendor_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Code"
          },
          "feedlot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedlot"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "lot_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Number"
          },
          "head_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Count"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/LineItemResponse"
            },
            "type": "array",
            "title": "Line Items",
            "default": []
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "extracted"
          },
          "validation_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Validation Warnings",
            "default": []
          },
          "erp_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erp Document Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "package_id",
          "invoice_number"
        ],
        "title": "InvoiceResponse",
        "description": "Invoice response."
      },
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "ready",
          "review",
          "blocked",
          "processing"
        ],
        "title": "InvoiceStatus",
        "description": "Invoice status values."
      },
      "InvoiceSummary": {
        "properties": {
          "invoice_id": {
            "type": "string",
            "title": "Invoice Id",
            "description": "Invoice identifier (e.g., INV-13304)"
          },
          "lot_number": {
            "type": "string",
            "title": "Lot Number",
            "description": "Lot identifier"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "description": "Invoice total amount"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus",
            "description": "Current invoice status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Reason if status is review/blocked"
          },
          "feed_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feed Type",
            "description": "Primary feed type"
          },
          "head_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Count",
            "description": "Number of head"
          },
          "days_on_feed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days On Feed",
            "description": "Days on feed"
          },
          "cost_per_head": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Head",
            "description": "Cost per head"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "AI extraction confidence (0-100)"
          }
        },
        "type": "object",
        "required": [
          "invoice_id",
          "lot_number",
          "amount",
          "status"
        ],
        "title": "InvoiceSummary",
        "description": "Summary of an invoice for list display within a package."
      },
...
    }
  }
}